Enhance rendering of code samples
**Code beautification**
Configuration for js-beautify is added in order to reflect our own .editorconfig (indent size of 2 spaces). Also, `no-preserve-newlines` is added for trimming newlines. When we use a lot of statements in the nunjucks file of a component, the output will have a lot of unnecessary newlines.

**Code highlighting**
I got fed up with the flash of unstyled code ;) Highlight.js is used for the code sample and will be loaded on every component detail page. There’s no need for parsing the sample code after loading, since it’s not displayed dynamically. When the code of a component gets updated, the page will be rendered again. So why not parse the code at the same time?

**Summary of changes**
- Removed static file for highlight.js
- Added highlight.js as a dependency
- Added configuration for js-beautify
- Updated to js-beautify 1.9.0