Remove Moment.js in favour of plain JS (#133)
I’m removing the dependency Moment.js and Moment-Timezone, because we don’t need it. We can use the Intl.DateTimeFormat object for quite a while now in Node.js. This change will save some disk space.

For delivery purposes, we show when the library has last been updated. This can be particularly useful for projects where we work with clients or colleagues that are in a different time zone. For extra clarity, the format will change from a mixed English/Dutch locale to British. `Last updated: 10-12-2020 13:37:00 CET` will become `Last updated: 10 December 2020 13:37:00 CET` with my change.

Configuration is available for people who need to set the correct time zone they’re working form, or when they want to change the format. In addition, configuration options are added as fallback for Node < 13.