fix(docs): support for Node v15+ (#143)
In PR #133 I’ve added redundant configuration for the date style. This was to provide a fallback to Node < 13. As of Node 15 this method of providing redundant config is not permitted anymore.

In this PR I’m removing the fallback in favour of clean code; making the Boilerplate more future proof.